% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bqc__norm__prob}
\alias{bqc__norm__prob}
\title{Evaluates whether \code{x} is a bare list}
\usage{
bqc__norm__prob(mu, sigma, a, b)
}
\arguments{
\item{mu}{\code{double}. The mean(s).}

\item{sigma}{\code{double}. The standard deviation(s).}

\item{a}{\code{numeric}. The lower bound(s).}

\item{b}{\code{numeric}. The upper bound(s).}

\item{x}{Any.}
}
\value{
\code{logical}.

\code{double}. Probability value(s).
}
\description{
Calculates the probability mass of a normal random variable over an interval.
}
\keyword{bqc__norm__prob()}
\keyword{internal}
